Generate arc geometry for a neighborhood glyph

Compute arc-level geometry used to render region- or cell-type–specific arcs in a cell neighborhood glyph.

Usage

generateArcData(
  node_data,
  arc_center,
  arc_radius = c(1, 0.9),
  arc_scale = c(0.7, 1),
  arc_angle = c(3 * pi/10, 7 * pi/10),
  arc_length_out = 100
)

Arguments

node_data

A data.frame of node-level information, typically generated by generateNodeData().

arc_center

A numeric vector of length 2 specifying the (x, y) center of the glyph.

arc_radius

A numeric vector specifying outer and inner arc radii. Default is c(1, 0.9).

arc_scale

A numeric vector of length 2 specifying anisotropic scaling factors applied to arc geometry in the x and y directions. The first value corresponds to the x-axis scale, and the second to the y-axis scale. Default is c(0.7, 1).

arc_angle

A numeric vector of length 2 specifying the start and end angles (in radians) of the arc. Default is c(3*pi/10, 7*pi/10).

arc_length_out

Integer specifying the number of points used to approximate each arc.

Value

A list containing arc geometry.